backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-jwt-extended==4.5.2
cryptography==41.0.3
pycryptodome==3.18.0
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.3.11
pinecone-client==2.2.4
bcrypt==4.0.1
sqlalchemy==2.0.19
flask-sqlalchemy==3.0.5
marshmallow==3.20.1
pytest==7.4.0
pytest-flask==1.2.0
gunicorn==21.2.0
